main-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.5.2

# MongoDB async driver
motor==3.6.0

# HTTP client for AI service communication
httpx==0.27.2

# Socket.IO
python-socketio==5.11.4


